+++ apps/wallet-mobile/src/features/Receive/common/ShareDetailsCard/useLastDateAddressUsed.ts
@@ -1,0 +1,20 @@
+import {useMemo} from 'react'
+
+import {useSelectedWallet} from '../../../../SelectedWallet'
+
+export const useLastDateAddressUsed = (address: string) => {
+  const wallet = useSelectedWallet()
+  const lastUsed = useMemo(
+    () =>
+      Object.values(wallet.transactions).reduce((currentLast, tx) => {
+        const {inputs, outputs} = tx
+        const isRelevant = inputs.some((v) => address === v.address) || outputs.some((v) => address === v.address)
+        if (!isRelevant) return currentLast
+        const lastUpdatedAt = new Date(tx.lastUpdatedAt).getTime()
+
+        return Math.max(currentLast, lastUpdatedAt)
+      }, 0),
+    [address, wallet.transactions],
+  )
+  return lastUsed
+}
